feat: add CheckboxDropdown component and integrate it into SearchBar
